+++ art-community-platform/backend/app/api/helpers/tag_helpers.py
@@ -1,4 +1,7 @@
-from ..models import User, ArtItem, Tag
+from ..models.user import User
+from ..models.art_item import ArtItem
+from ..models.comment import Comment
+from ..models.tag import Tag
 
 
 def get_tag_by_id_helper(tag_id):
