+++ art-community-platform/backend/app/api/models/exhibition.py
@@ -1,0 +1,3 @@
+from django.db import models
+from django.contrib.postgres.fields import ArrayField
+
