+++ art-community-platform/backend/app/api/models/tag.py
@@ -1,0 +1,7 @@
+from django.db import models
+from django.contrib.postgres.fields import ArrayField
+
+
+class Tag(models.Model):
+    text = models.CharField(max_length=100)
+
