+++ art-community-platform/backend/app/api/models/user.py
@@ -21,23 +21,3 @@ class User(models.Model):
 
     def __str__(self):
         self.email + " " + self.token
-
-
-class ArtItem(models.Model):
-    owner_id = models.IntegerField()
-    img_url = models.URLField()
-    description = models.CharField(max_length=512)
-    date = models.DateField()
-    tags = ArrayField(models.CharField(max_length=100), null=True)
-    comments = ArrayField(models.IntegerField(), null=True)
-    favourites = ArrayField(models.IntegerField(), null=True)
-
-
-class Comment(models.Model):
-    owner_id = models.IntegerField()
-    text = models.CharField(max_length=1024)
-    date = models.DateField()
-
-
-class Tag(models.Model):
-    text = models.CharField(max_length=100)
