+++ art-community-platform/backend/app/api/serializers/serializers.py
@@ -1,5 +1,8 @@
 from rest_framework import serializers
-from .models import User, ArtItem, Comment, Tag
+from ..models.user import User
+from ..models.art_item import ArtItem
+from ..models.comment import Comment
+from ..models.tag import Tag
 
 
 class UserSerializer(serializers.ModelSerializer):
