+++ art-community-platform/backend/app/api/tests/test_urls.py
