+++ art-community-platform/backend/app/app/urls.py
@@ -20,8 +20,8 @@ from django.conf import settings
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('user/', include('user.urls')),
-    path('api/v1/', include('user.urls')),
+    path('api/', include('api.urls')),
+    path('api/v1/', include('api.urls')),
 ]
 
 if settings.DEBUG:
